+++ core/controllers/question_editor_test.py
@@ -1,4 +1,4 @@
-# Copyright 2017 The Oppia Authors. All Rights Reserved.
+# Copyright 2018 The Oppia Authors. All Rights Reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
